requirements.txt
astroid==2.4.2
async-generator==1.10
async-lru==1.0.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pyaes==1.6.1
pylint==2.5.3
pymongo==3.11.0
https://github.com/pyrogram/pyrogram/archive/asyncio.zip
PySocks==1.7.1
python-dotenv==0.14.0
six==1.15.0
toml==0.10.1
wrapt==1.12.1
